fix: force retry to open database with legacy key method
When opening old databases (created with version 3.4.0) with the new native "key" methods fails, because of a fix done in #######
To fix this issue, when first failing to open the database, the system will try to open the database using the old key method. If it succeeds to open the DB then it will change the password to the same password but now using the new key methods. This way, this error never occurs again to that instance of the DB.